include: new header for fundamental pointer types
This header is holding forward declarations that are needed in many places.
It's for helping us to unclutter the include files a little bit.
