+++ include/input.h
@@ -47,6 +47,8 @@ SOFTWARE.
 #ifndef INPUT_H
 #define INPUT_H
 
+#include "xlibre_ptrtypes.h"
+
 #include "misc.h"
 #include "screenint.h"
 #include <X11/Xmd.h>
