+++ include/randrstr.h
@@ -30,6 +30,8 @@
 
 #include <X11/X.h>
 #include <X11/Xproto.h>
+
+#include "xlibre_ptrtypes.h"
 #include "misc.h"
 #include "os.h"
 #include "dixstruct.h"
